fix 0.3 User Info connect db
