+++ app/src/views/UserInfo/userInfoForm.tsx
@@ -1,20 +1,7 @@
-import {
-  Avatar,
-  Box,
-  Button,
-  Container,
-  Grid,
-  IconButton,
-  Menu,
-  MenuItem,
-  Stack,
-  TextField,
-} from "@mui/material";
-import { SubmitHandler, useForm } from "react-hook-form";
+import { Box, Button, Container, Grid, Stack, TextField } from "@mui/material";
+import { SubmitHandler, useForm, FormProvider } from "react-hook-form";
 import Typography from "@mui/material/Typography";
-import { Delete, Photo } from "@mui/icons-material";
-import PopupState, { bindTrigger, bindMenu } from "material-ui-popup-state";
-import { useEffect, useState } from "react";
+import React from "react";
 import { useAuth } from "../../hooks/useAuth";
 import db from "../../services/firebase/firestore";
 import { doc, setDoc } from "firebase/firestore";
@@ -22,10 +9,7 @@ import { UserInfoDB } from "../../../../typescript-types/db.types";
 import { ref, uploadBytes, getDownloadURL } from "firebase/storage";
 import { storage } from "../../services/firebase";
 import { updateProfile } from "firebase/auth";
-
-interface Event<T = EventTarget> {
-  target: T;
-}
+import { AvatarInput } from "./avatar.input";
 
 export interface Inputs {
   avatar?: FileList;
@@ -42,64 +26,39 @@ interface preloadedValuesProps {
 }
 
 export function UserInfoForm({ preloadedValues }: preloadedValuesProps): JSX.Element {
-  const [avatar, setAvatar] = useState("");
   const { uid, user } = useAuth();
 
-  useEffect(() => {
-    setAvatar(user?.photoURL || "");
-  }, []);
-
-  const {
-    register,
-    handleSubmit,
-    setValue,
-    formState: { errors },
-  } = useForm<Inputs>({
+  const methods = useForm<Inputs>({
     defaultValues: preloadedValues,
   });
 
   const avatarUpdate = (avatarToUpdate: File) => {
     if (user) {
-      if (avatar) {
-        const storageRef = ref(storage, `images/avatars/${uid}`);
-        uploadBytes(storageRef, avatarToUpdate).then((uploadTask) => {
-          console.log("Uploaded a blob or file!");
-          getDownloadURL(uploadTask.ref).then((downloadURL) => {
-            console.log("File available at", downloadURL);
-            updateProfile(user, {
-              photoURL: downloadURL,
-            }).then(() => {
-              console.log("update avatarToUpdate");
-            });
+      // if () {
+      const storageRef = ref(storage, `images/avatars/${uid}`);
+      uploadBytes(storageRef, avatarToUpdate).then((uploadTask) => {
+        console.log("Uploaded a blob or file!");
+        getDownloadURL(uploadTask.ref).then((downloadURL) => {
+          console.log("File available at", downloadURL);
+          updateProfile(user, {
+            photoURL: downloadURL,
+          }).then(() => {
+            console.log("update avatarToUpdate");
           });
         });
-      } else {
-        updateProfile(user, {
-          photoURL: "",
-        }).then(() => {
-          console.log("delete avatar");
-        });
-      }
+      });
+      // }
+      // else {
+      //   updateProfile(user, {
+      //     photoURL: "",
+      //   }).then(() => {
+      //     console.log("delete avatar");
+      //   });
+      // }
     }
   };
 
-  const helper = errors.lastName ? "Error message" : "";
-
-  const DeletePhoto = () => {
-    const file = new File([""], "deletedAvatar", { type: "image/png" });
-    const dt = new DataTransfer();
-    dt.items.add(file);
-    const fileList = dt.files;
-    setValue("avatar", fileList);
-    setAvatar("");
-  };
-
-  const showPreview = (event: Event<HTMLInputElement>) => {
-    if (event.target && event.target.files && event.target.files.length > 0) {
-      const src = URL.createObjectURL(event.target.files[0]);
-      setAvatar(src);
-    }
-  };
+  const helper = methods.formState.errors.lastName ? "Error message" : "";
 
   const onSubmit: SubmitHandler<Inputs> = async (data) => {
     console.log("отправка");
@@ -118,106 +77,69 @@ export function UserInfoForm({ preloadedValues }: preloadedValuesProps): JSX.Ele
 
   return (
     <Container sx={{ display: "flex", justifyContent: "center" }}>
-      <form onSubmit={handleSubmit(onSubmit)}>
-        <Stack spacing={2} width={508} padding={10} sx={{ paddingTop: "48px" }}>
-          <Box sx={{ paddingBottom: "24px" }}>
-            <Typography variant="h2" component="h2">
-              User Info
-            </Typography>
-          </Box>
-          <Grid container spacing={0}>
-            <Grid item xs={4}>
-              <PopupState variant="popover" popupId="demo-popup-menu">
-                {(popupState) => (
-                  <>
-                    <TextField
-                      type={"file"}
-                      sx={{ display: "none" }}
-                      id="select-image"
-                      {...register("avatar", { required: false, onChange: (e) => showPreview(e) })}
-                    />
-                    <label htmlFor={avatar ? "" : "select-image"}>
-                      <IconButton
-                        sx={{ height: 160, width: 160 }}
-                        component="span"
-                        {...bindTrigger(popupState)}
-                      >
-                        <Avatar sx={{ width: 160, height: 160 }} src={avatar} />
-                      </IconButton>
-                    </label>
-                    {avatar && (
-                      <Menu {...bindMenu(popupState)}>
-                        <label htmlFor="select-image">
-                          <MenuItem onClick={popupState.close}>
-                            <Photo sx={{ paddingRight: "5px" }} />
-                            Change Photo
-                          </MenuItem>
-                        </label>
-                        <MenuItem
-                          onClick={() => {
-                            DeletePhoto();
-                            popupState.close();
-                          }}
-                        >
-                          <Delete sx={{ paddingRight: "5px" }} />
-                          Delete Photo
-                        </MenuItem>
-                      </Menu>
-                    )}
-                  </>
-                )}
-              </PopupState>
+      <FormProvider {...methods}>
+        <form onSubmit={methods.handleSubmit(onSubmit)}>
+          <Stack spacing={2} width={508} padding={10} sx={{ paddingTop: "48px" }}>
+            <Box sx={{ paddingBottom: "24px" }}>
+              <Typography variant="h2" component="h2">
+                User Info
+              </Typography>
+            </Box>
+            <Grid container spacing={0}>
+              <Grid item xs={4}>
+                {user && <AvatarInput {...user} />}
+              </Grid>
+              <Grid item xs={8}>
+                <Stack spacing={"24px"} width={316} paddingLeft={"22.66px"}>
+                  <TextField
+                    label={"First Name"}
+                    error={!!methods.formState.errors.firstName}
+                    helperText={helper}
+                    placeholder={"Enter your first name"}
+                    {...methods.register("firstName", { required: true })}
+                  />
+                  <TextField
+                    label={"Last Name"}
+                    {...methods.register("lastName", { required: true })}
+                    placeholder={"Enter your last name"}
+                    error={!!methods.formState.errors.lastName}
+                    helperText={helper}
+                  />
+                </Stack>
+              </Grid>
             </Grid>
-            <Grid item xs={8}>
-              <Stack spacing={"24px"} width={316} paddingLeft={"22.66px"}>
-                <TextField
-                  label={"First Name"}
-                  error={!!errors.firstName}
-                  helperText={helper}
-                  placeholder={"Enter your first name"}
-                  {...register("firstName", { required: true })}
-                />
-                <TextField
-                  label={"Last Name"}
-                  {...register("lastName", { required: true })}
-                  placeholder={"Enter your last name"}
-                  error={!!errors.lastName}
-                  helperText={helper}
-                />
-              </Stack>
-            </Grid>
-          </Grid>
-          <TextField
-            label={"Email"}
-            {...register("email", { required: false })}
-            disabled={true}
-            placeholder={"Placeholder"}
-          />
-          <TextField
-            multiline
-            rows={4}
-            label={"About"}
-            {...register("about", { required: false })}
-            placeholder={"Provide short description about yourself"}
-          />
-          <TextField
-            label={"Phone"}
-            {...register("phone", { required: false })}
-            placeholder={"+XXX (XX) XXX-XX-XX"}
-          />
-          <TextField
-            label={"LinkedIn"}
-            {...register("linkedIn", { required: false })}
-            placeholder={"Enter your LinkedIn URL"}
-          />
-          <Stack paddingTop={"40px"} spacing={2} direction={"row"}>
-            <Button variant={"contained"} type="submit">
-              Save Changes
-            </Button>
-            <Button variant={"outlined"}>Cancel</Button>
+            <TextField
+              label={"Email"}
+              {...methods.register("email", { required: false })}
+              disabled={true}
+              placeholder={"Placeholder"}
+            />
+            <TextField
+              multiline
+              rows={4}
+              label={"About"}
+              {...methods.register("about", { required: false })}
+              placeholder={"Provide short description about yourself"}
+            />
+            <TextField
+              label={"Phone"}
+              {...methods.register("phone", { required: false })}
+              placeholder={"+XXX (XX) XXX-XX-XX"}
+            />
+            <TextField
+              label={"LinkedIn"}
+              {...methods.register("linkedIn", { required: false })}
+              placeholder={"Enter your LinkedIn URL"}
+            />
+            <Stack paddingTop={"40px"} spacing={2} direction={"row"}>
+              <Button variant={"contained"} type="submit">
+                Save Changes
+              </Button>
+              <Button variant={"outlined"}>Cancel</Button>
+            </Stack>
           </Stack>
-        </Stack>
-      </form>
+        </form>
+      </FormProvider>
     </Container>
   );
 }
